Task_04/study-assistant/requirements.txt
openai-agents==0.6.1
pypdf==6.4.0
streamlit==1.51.0
python-dotenv==1.2.1
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.3.0
